simulator_helper

+++ app/controllers/simulator_controller.rb
@@ -1,4 +1,5 @@
 class SimulatorController < ApplicationController
+  include SimulatorHelper
 
   before_action :authenticate_user!, only: [:create, :update, :edit,:update_image]
   before_action :set_project, only: [:show, :embed, :embed, :update, :edit, :get_data,:update_image]
@@ -41,14 +42,17 @@ class SimulatorController < ApplicationController
 
   def update
     @project.data = params[:data]
-    data_url = params[:image]
-    jpeg      = Base64.decode64(data_url['data:image/jpeg;base64,'.length .. -1])
-    image_file = File.new("preview_#{Time.now()}.jpeg", "wb")
-    image_file.write(jpeg)
+
+    image_file = return_image_file(params[:image])
+
     @project.image_preview = image_file
     @project.name = params[:name]
     @project.save
-    File.delete(image_file)
+
+    if check_to_delete(params[:image])
+      File.delete(image_file)
+    end
+
     render plain: "success"
   end
 
@@ -59,23 +63,12 @@ class SimulatorController < ApplicationController
     @project.name = params[:name]
     @project.author = current_user
 
-    data_url = params[:image]
-
-    str = data_url['data:image/jpeg;base64,'.length .. -1]
-    if str.to_s.empty?
-      path = Rails.root.join("app/assets/images/empty_project/no_preview.png")
-      image_file = File.open(path, "rb")
-
-    else
-      jpeg       = Base64.decode64(str)
-      image_file = File.new("preview_#{Time.now()}.jpeg", "wb")
-      image_file.write(jpeg)
-    end
+    image_file = return_image_file(params[:image])
 
     @project.image_preview = image_file
     @project.save
 
-    if !str.to_s.empty?
+    if check_to_delete(params[:image])
       File.delete(image_file)
     end
